Cleaned up controllers and set up predefined functions to interact with the database for now just working on being able to console.log all the needed functions.
