+++ controllers/appointments.js
@@ -1,0 +1,25 @@
+// *************************************************************************************************
+//                              controllers/appointments.js  
+//      This file will contain all the functions required to update or make appointments 
+//                          
+//
+//     
+// *************************************************************************************************
+
+
+var db = require("../models");
+
+
+var exports = module.exports = {};
+
+    //This function will get the required information from the data object argument then update the Appointment, and then will pass back wether it failed or succeeded through the callback
+    exports.update = function(data,cb){
+
+    };
+
+    //This function will gather the information needed from the data object arguement to create a new Appointment, and then will pass back wether it failed or succeeded through the callback
+
+    exports.newAppointment = function(data,cb){
+
+    }
+
