+++ controllers/rendering-controller.js
@@ -1,9 +1,37 @@
+//
+// *************************************************************************************************
+//                              controllers/rendering-controller.js  
+//
+//      THIS PAGE WILL HANDLE ANYTHING DEALING WITH GATHERING STATIC INFORMATION AND RENDERING 
+//                              IT THROUGH TO THE HANDLEBARS
+//
+// *************************************************************************************************
+
 var db = require("../models");
-var getCustomer = require('./customer-api.js');
-var getService = require('./services-api-routes.js')
+// Including the modules for Customers
+var getCustomer = require('./customer.js');
+
+//Including the modules for Services
+var getService = require('./services.js')
+
+//Including the modules for Providers
+var getProviders = require("./provider.js");
+
+//Including the modules for Appointments
+var getAppointments = require('./appointments.js');
+
+//Including the modules for Schedules
+var getSchedules = require('./schedules.js')
+
+
 var exports = module.exports = {};
+
+
 //Sets viewBuilder as global so it can be access by all routes
 var viewBuilder = {};
+
+// This Function will handle logic for when users go to the path '/'
+// This will figure out if there's a user already logged in by checking if(req.user) and redirect back to /home + the other params. If there's no user just redirect to /home
 exports.Welcome = function(req,res, next){
     if(req.user){
         res.redirect('/home/' + req.user.userType + '/' + req.user.id)
@@ -13,17 +41,19 @@ exports.Welcome = function(req,res, next){
     
 }
 
-exports.home = function(req, res){
 
+// This function will be the main function that buils the viewBuilder containing all the stuff we're going to need to render in views and will be created at the '/hjome' path
+exports.home = function(req, res){
+    //If there's no user logged in this will be the basic view builder
     viewBuilder = {
         LoggedIn : false,
         messageLogIn : req.flash('logInMessage'), 
         messageSignIn: req.flash('signUpMessage'),
         messagesettingsMessage: req.flash('settingsMessage'),
         admin: false
-    }
+    };
 
-
+    // If there is a user logged all of this will be added to viewBuilder
     if(req.user){
         viewBuilder.LoggedIn = true;
         viewBuilder.userId = req.user.id;
@@ -41,6 +71,7 @@ exports.home = function(req, res){
         viewBuilder.scheduleRoute = '/schedule/'+req.user.userType+'/'+req.user.id;
         viewBuilder.updateRoute = "/update/" + req.user.userType + "/" + req.user.id
 
+        //If a user is a Admin set the admin equal to true
         if(req.user.userType === 'admin'){
             viewBuilder.admin = true;
         }
@@ -52,6 +83,7 @@ exports.home = function(req, res){
 
 }
 
+// Handles for when user goes to the path '/about'
 exports.about = function (req, res) {
     res.render('about', viewBuilder);
 };
@@ -59,7 +91,9 @@ exports.about = function (req, res) {
 
 
 
+// Handles when users go to providers page!!!***THIS FUNCTION STILL NEEDS MORE INFORMATION TO BE GATHERED BEFORE IT CAN RENDER ALL STATIC INFORMATION ON THIS PAGE
 exports.provider = function (req, res) {
+
 
     getCustomer.AllInfo(function(err,clients){
         console.log(clients);
@@ -72,8 +106,13 @@ exports.provider = function (req, res) {
 };
 
 
+// Handles when users go to services page!!!***THIS FUNCTION STILL NEEDS MORE INFORMATION TO BE GATHERED BEFORE IT CAN RENDER ALL STATIC INFORMATION ON THIS PAGE
+
 exports.service = function (req, res) {
 
+    //
+    //This is were we get all services information
+    //
     getService.AllServices(function(err,services){
 
         console.log(services);
