+++ controllers/schedule.js
@@ -1,0 +1,21 @@
+//
+// *************************************************************************************************
+//                                      controllers/schedule.js  
+//
+//      THIS PAGE WILL HANDLE ANYTHING DEALING WITH GATHERING STATIC INFORMATION AND RENDERING 
+//                              IT THROUGH TO THE HANDLEBARS
+//
+// *************************************************************************************************
+
+
+//Importing our database
+var db = require("../models");
+
+var exports = module.exports = {};
+
+    //This functions is.... uhh getting all Schedules with joined Appointments then right joining it with Providers o.O? Good luck.. It passes the information back through a callback
+    exports.getWithAppRProv = function(cb){
+
+    }
+
+
