+++ controllers/services.js
@@ -1,5 +1,6 @@
 // *********************************************************************************
-// api-routes.js - this file offers a set of routes for displaying and saving data to the db
+//                      controllers/services-api-routes.js 
+//  This file offers a set of function for retrieving and updating data in services
 // *********************************************************************************
 
 // Dependencies
@@ -9,11 +10,10 @@ var db = require("../models");
 
 var exports = module.exports = {};
 
+  // This functions gets all the services from the database and passes back through a callback
   exports.AllServices = function(cb){
 
-      db.Services.findAll(
-        // include: [db.Providers]
-      ).then(function (dbService) {
+      db.Services.findAll().then(function (dbService) {
         if(dbService.length > 0){
           var services = [];
           for(var i = 0; i < dbService.length; i++){
@@ -32,5 +32,18 @@ var exports = module.exports = {};
   
         
       });
+
+  }
+
+
+  // This function will retrieve the needed information through the data arguement to update a Service from the database
+  exports.update = function(data,cb){
+
+  }
+
+              //IF WE HAVE TIME
+  // This function will get the needed information about a Service and remove it from the database 
+
+  exports.removeService = function(data,cb){
 
   }
