+++ server.js
@@ -63,7 +63,7 @@ app.use(session({
 
 // Routes
 // =============================================================
-// require("./routes/api-routes.js")(app);
+require("./routes/api-routes.js")(app);
 require("./routes/html-routes.js")(app);
 require('./routes/userAuth.js')(app,passport)
